move test folders to correct folder
